+++ CHANGELOG.md
@@ -2,7 +2,5 @@
 This is a bug fix release for various issues discovered after the release v0.3.0. Also new features have been added.
 
 ## What's Changed
+* FEATURE: each project has a dedicated drawer for todos.
 * FIXED: fixed bug related to importlib.util error.
-* FEATURE: added download button on the image viewer.
-* OPTIMIZED: projects have a dedicated uuid when created which is used instead of the id.
-* OPTIMIZED: user can create a `typst.preamble` file at the root directory of the project. Typst will use this file when rendering chapter level PDF.